WT-7782 Separate implementation from headers in cppsuite test framework (#6772)
